@@ -19,15 +19,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.4",
-    "@angular/common": "13.3.4",
-    "@angular/compiler": "13.3.4",
-    "@angular/core": "13.3.4",
-    "@angular/forms": "13.3.4",
-    "@angular/platform-browser": "13.3.4",
-    "@angular/platform-browser-dynamic": "13.3.4",
-    "@angular/platform-server": "13.3.4",
-    "@angular/router": "13.3.4",
+    "@angular/animations": "14.3.0",
+    "@angular/common": "14.3.0",
+    "@angular/compiler": "14.3.0",
+    "@angular/core": "14.3.0",
+    "@angular/forms": "14.3.0",
+    "@angular/platform-browser": "14.3.0",
+    "@angular/platform-browser-dynamic": "14.3.0",
+    "@angular/platform-server": "14.3.0",
+    "@angular/router": "14.3.0",
     "@delon/abc": "^13.4.2",
     "@delon/acl": "^13.4.2",
     "@delon/auth": "^13.4.2",
@@ -45,12 +45,12 @@
     "core-js": "^3.3.3",
     "date-fns": "^2.12.0",
     "jquery": "3.5.0",
-    "ng-zorro-antd": "^13.1.1",
+    "ng-zorro-antd": "^14.3.0",
     "ngx-tinymce": "^13.0.0",
     "ngx-ueditor": "^13.0.0",
     "oidc-client": "^1.9.1",
     "popper.js": "^1.16.0",
-    "rxjs": "^6.5.3",
+    "rxjs": "~7.8.0",
     "screenfull": "^5.0.2",
     "tslib": "^2.0.0",
     "wangeditor": "^4.7.4",
@@ -58,10 +58,10 @@
     "ajv-formats": "^2.1.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.3",
-    "@angular/cli": "^13.3.3",
-    "@angular/compiler-cli": "13.3.4",
-    "@angular/language-service": "~13.3.0",
+    "@angular-devkit/build-angular": "^14.2.12",
+    "@angular/cli": "^14.2.12",
+    "@angular/compiler-cli": "14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@delon/testing": "^13.4.2",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.8",
@@ -81,14 +81,14 @@
     "ng-alain-codelyzer": "^0.0.1",
     "typescript": "~4.6.3",
     "xlsx": "^0.17.0",
-    "@angular-eslint/builder": "~13.1.0",
-    "@angular-eslint/eslint-plugin": "~13.1.0",
-    "@angular-eslint/eslint-plugin-template": "~13.1.0",
-    "@angular-eslint/schematics": "~13.1.0",
-    "@angular-eslint/template-parser": "~13.1.0",
-    "@typescript-eslint/eslint-plugin": "~5.15.0",
-    "@typescript-eslint/parser": "~5.15.0",
-    "eslint": "^8.11.0",
+    "@angular-eslint/builder": "~14.4.0",
+    "@angular-eslint/eslint-plugin": "~14.4.0",
+    "@angular-eslint/eslint-plugin-template": "~14.4.0",
+    "@angular-eslint/schematics": "~14.4.0",
+    "@angular-eslint/template-parser": "~14.4.0",
+    "@typescript-eslint/eslint-plugin": "^5.36.2",
+    "@typescript-eslint/parser": "^5.36.2",
+    "eslint": "^8.23.0",
     "eslint-config-prettier": "^2.6.0",
     "eslint-plugin-import": "~2.25.4",
     "eslint-plugin-jsdoc": "~38.0.4",
@@ -104,7 +104,6 @@
     "tslint": "~6.1.0"
   },
   "husky": {
-    "hooks": {
-    }
+    "hooks": {}
   }
 }
